% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_luf16.R
\name{read_luf16}
\alias{read_luf16}
\title{Read ListUserFile16}
\usage{
read_luf16(rep_file)
}
\arguments{
\item{rep_file}{Path and filename of ListUserFile16}
}
\description{
Read LSSS-report ListUserFile16 giving 3D (layered) NASC by species.
}
