% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_luf3.R
\name{read_luf3}
\alias{read_luf3}
\title{Read ListUserFile03}
\usage{
read_luf3(rep_file)
}
\arguments{
\item{rep_file}{Path and filename of ListUserFile03.}
}
\description{
Read LSSS-report ListUserFile03 giving 2D (water column) NASC by species.
}
